fix: handle webpack config exported as object
