+++ packages/core/config/webpack.production.js
@@ -1,5 +1,6 @@
 const webpackConfig = require('./webpack.common')
-const { mergeUserWebpackConfig } = require('@redwoodjs/internal')
+const { mergeUserWebpackConfig } = require('./utils')
+
 const baseConfig = webpackConfig('production')
 
 module.exports = mergeUserWebpackConfig('production', baseConfig)
